refine run_inference

+++ mri/model_mri.py
@@ -53,7 +53,7 @@ class STCNNT_MRI(STCNNT_Task_Base):
             
         if config.backbone == "hrnet":
             
-            hrnet_C_out = config.backbone_hrnet.C * sum([np.power(2, k) for k in range(config.backbone_hrnet.num_resolution_levels)])
+            hrnet_C_out = int(config.backbone_hrnet.C * sum([np.power(2, k) for k in range(config.backbone_hrnet.num_resolution_levels)]))
             
             self.pre = nn.Identity()
             self.backbone = STCNNT_HRnet(config=config)
@@ -65,7 +65,7 @@ class STCNNT_MRI(STCNNT_Task_Base):
             self.post = Conv2DExt(config.backbone_unet.C, config.C_out, kernel_size=config.kernel_size, stride=config.stride, padding=config.padding, bias=True)
             
         if config.backbone == "LLM":
-            output_C = np.power(2, config.backbone_LLM.num_stages-2) if config.backbone_LLM.num_stages>2 else config.backbone_LLM.C
+            output_C = int(np.power(2, config.backbone_LLM.num_stages-2)) if config.backbone_LLM.num_stages>2 else config.backbone_LLM.C
             self.pre = nn.Identity()
             self.backbone = STCNNT_LLMnet(config=config) 
             self.post = Conv2DExt(output_C,config.C_out, \
